feat: rename mmap feature to mman
