+++ Cargo.toml
@@ -58,7 +58,8 @@ gem-net = ["tcp", "dep:tock-registers"]
 idle-poll = []
 log-target = []
 net = []
-mmap = []
+mman = []
+mmap = ["mman"] # Deprecated in favor of mman
 newlib = []
 nostd = []
 pci = ["virtio/pci"]
