+++ src/syscalls/mod.rs
@@ -35,8 +35,8 @@ mod condvar;
 mod entropy;
 mod futex;
 pub(crate) mod interfaces;
-#[cfg(feature = "mmap")]
-mod mmap;
+#[cfg(feature = "mman")]
+mod mman;
 mod processor;
 #[cfg(feature = "newlib")]
 mod recmutex;
